+++ apiserver/common/export_test.go
@@ -4,10 +4,11 @@
 package common
 
 var (
-	ValidateNewFacade = validateNewFacade
-	WrapNewFacade     = wrapNewFacade
-	NilFacadeRecord   = facadeRecord{}
-	EnvtoolsFindTools = &envtoolsFindTools
+	MachineJobFromParams = machineJobFromParams
+	ValidateNewFacade    = validateNewFacade
+	WrapNewFacade        = wrapNewFacade
+	NilFacadeRecord      = facadeRecord{}
+	EnvtoolsFindTools    = &envtoolsFindTools
 )
 
 type Patcher interface {
